requirements.txt
PyQt5
groq
keyboard
pystray
Pillow
pywin32
